+++ documentation/changelog.md
@@ -13,6 +13,17 @@ This is the changelog for `casual` and all changes are listed in this document.
    `transaction --list-instances` now lists all instances of all resources, _internal_
    and _external_. This makes it consistent with _service --list-instances_. 
 
+## [1.6.21] - 2024-10-09
+### Fixes
+- documentation: queue.forward.group.(services|queues).alias
+- transaction: error log when resource fail (#428)
+
+## [1.6.20] - 2024-09-18
+### Fixes
+- build: handle nginx build on rhel (#381)
+- service: fix invisible route being created when updating service config
+- queue: enqueue/dequeue to multiple queues in example server
+
 ## [1.6.19] - 2024-04-25
 ### Fixes
 - http: improved error logging in http-outbound (#367)